// src/lib/tchart/types.d.ts
export type TChartSettings = {
	isIE: boolean;
	isEdge: boolean;
	dpi: number;
	darkMode: boolean;
	ALL_LABEL: string;
	Y_AXIS_RANGE: number;
	PADD: [number, number, number, number];
	PADD_MINI: [number, number, number, number];
	PADD_MINI_BAR: [number, number, number, number];
	PADD_MINI_AREA: [number, number, number, number];
	Y_LABELS_WIDTH: number;
	X_LABELS_HEIGHT: number;
	DATES_HEIGHT: number;
	DATES_WIDTH: number;
	DATES_SIDE: "left" | "right";
	DATES_WEIGHT: "bold" | "normal";
	DATES_FONT_SIZE?: number;
	ZOOM_TEXT: string;
	MINI_GRAPH_HEIGHT: number;
	MINI_GRAPH_TOP: number;
	MINI_GRAPH_BOTTOM: number;
	FADE_HEIGHT: number;
	PIE_RADIUS: number;
	FONT: {
		family: string;
		bold: string;
		normal: string;
	};
	COLORS: {
		primary?: string; // used in css
		secondary?: string; // used in css
		background: string;
		backgroundRgb: [number, number, number];
		text: string;
		dates: string;
		grid: string;
		axis: {
			x: string;
			y: string;
		};
		barsSelectionBackground: string;
		miniMask: string;
		miniFrame: string;
	};
};

export type TChartDataDetails = Partial<{
	maxXTickLength: number;
	x: number[];
	y: number[][];
	names: string[];
}> &
	Partial<
		Pick<
			TChartConstructorOptions["data"],
			| "subchart"
			| "types"
			| "columns"
			| "hidden"
			| "yTickFormatter"
			| "yTooltipFormatter"
			| "xTickFormatter"
			| "xTooltipFormatter"
			| "xRangeFormatter"
			| "getLabelDate"
			| "getLabelTime"
		>
	>;

export type TChartData = TChartDataDetails & {
	caption: string;
	detailsFunc: (x: number) => Promise<TChartConstructorOptions["data"]>;
	hasDetail: boolean;
	slave: boolean;
	yMinStep: number;
	xRangeFormatter: any;
	strokeWidth: number | "auto";
	tooltipOnHover: boolean;
	forceLegend: boolean;
	sideLegend: boolean;
	pieZoomRange: number;
	pieLabelsPercentages: {
		outboard: number;
		hoverOnly: number;
	};

	mainPeriodLen?: number;
	detailPeriodLen?: number;
	dates?: string[];
	datesShort?: string[];
	datesRange?: string[];
	ys?: {
		colors_d: [string, string, string];
		colors_n: [string, string, string];
		label: string;
		y: number[];
		tp: string;
		id: string;
		yFrom?: number[];
		outside?: boolean;
	}[];
	yIds?: any;
	saved?: any;
	master?: boolean;
	details?: TChartDataDetails;
};
// ...
export type TChartConstructorOptions = {
	container: HTMLElement;
	data: TChatOriginalData & {
		getLabelDate?: typeof getLabelDate;
		getLabelTime?: typeof getLabelTime;
		x_on_zoom?: TChartData["detailsFunc"];
		yMinStep?: number;
		tooltipOnHover?: boolean;
		forceLegend?: boolean;
		sideLegend?: boolean;
		pieZoomRange?: number;
		pieLabelsPercentages?: {
			outboard: number;
			hoverOnly: number;
		};
	};
	settings?: Partial<TChartSettings>;
	slave?: boolean;
};

export type TChatOriginalData = {
	colors: Record<string, string>;
	columns: number[][]; // in reality it will be ['x0', ...number]
	hidden: string[];
	names: Record<string, string>;
	strokeWidth?: number;
	subchart?: {
		show: boolean;
		defaultZoom: [number, number];
	};
	title: string;
	types: Record<string, string>;
	xRangeFormatter: string;
	xTickFormatter: string;
	xTooltipFormatter: string;
	yTickFormatter: string;
	yTooltipFormatter: string;
	y_scaled?: boolean;
};